# docs/openapi.json
{
  "components": {
...
    "schemas": {
...
      "CreateRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/CreateRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "contact_email": {
            "type": "string"
          },
          "first_station": {
            "$ref": "#/components/schemas/FirstStationRequest"
          },
          "legal_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "legal_name",
          "address",
          "contact_email",
          "timezone",
          "first_station"
        ],
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "location": {
            "description": "Where the error occurred, e.g. 'body.items[3].tags' or 'path.thing-id'",
            "type": "string"
          },
          "message": {
            "description": "Error message text",
            "type": "string"
          },
          "value": {
            "description": "The value at the given location"
          }
        },
        "type": "object"
      },
      "ErrorModel": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/ErrorModel.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "code": {
            "description": "Stable machine error code",
            "type": "string"
          },
          "detail": {
            "description": "A human-readable explanation specific to this occurrence of the problem.",
            "examples": [
              "Property foo is required but is missing."
            ],
            "type": "string"
          },
          "errors": {
            "description": "Optional list of individual error details",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "field_errors": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object"
          },
          "instance": {
            "description": "A URI reference that identifies the specific occurrence of the problem.",
            "examples": [
              "https://example.com/error-log/abc123"
            ],
            "format": "uri",
            "type": "string"
          },
          "message": {
            "description": "Safe user message",
            "type": "string"
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "description": "HTTP status code",
            "examples": [
              400
            ],
            "format": "int64",
            "type": "integer"
          },
          "title": {
            "description": "A short, human-readable summary of the problem type. This value should not change between occurrences of the error.",
            "examples": [
              "Bad Request"
            ],
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "description": "A URI reference to human-readable documentation for the error.",
            "examples": [
              "https://example.com/errors/example"
            ],
            "format": "uri",
            "type": "string"
          }
        },
        "type": "object"
      },
      "FirstStationRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "timezone"
        ],
        "type": "object"
      },
...
      "OrganizationDetail": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/OrganizationDetail.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "contact_email": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "legal_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "org_id",
          "name",
          "legal_name",
          "address",
          "contact_email",
          "timezone",
          "enabled",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
      "OrganizationUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/OrganizationUpdateRequest.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "contact_email": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "legal_name": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "sessionCookie": {
        "description": "httpOnly session cookie",
        "in": "cookie",
        "name": "pomkita_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "PomKita API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/organizations": {
      "get": {
        "operationId": "listOrganizations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrganizationDetail"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "List organizations",
        "x-permitted-roles": [
          "Owner",
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "organization_administration_forbidden"
          ]
        }
      },
      "post": {
        "operationId": "createOrganization",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "example": {
                  "summary": "Request example",
                  "value": {
                    "address": "Jakarta",
                    "contact_email": "admin@example.test",
                    "first_station": {
                      "name": "Main",
                      "timezone": "Asia/Jakarta"
                    },
                    "legal_name": "Pom Org PT",
                    "name": "Pom Org",
                    "timezone": "Asia/Jakarta"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDetail"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create an organization",
        "x-csrf-required": true,
        "x-permitted-roles": [
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "organization_administration_forbidden",
            "csrf_required"
          ],
          "422": [
            "first_station_required",
            "invalid_organization_request"
          ]
        }
      }
    },
    "/api/v1/organizations/{id}": {
      "get": {
        "operationId": "getOrganization",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDetail"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Read an organization",
        "x-permitted-roles": [
          "Owner",
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "organization_administration_forbidden"
          ],
          "404": [
            "organization_not_found"
          ]
        }
      },
      "patch": {
        "operationId": "updateOrganization",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "example": {
                  "summary": "Request example",
                  "value": {
                    "enabled": true,
                    "name": "Pom Org Updated"
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/OrganizationUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationDetail"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update an organization",
        "x-csrf-required": true,
        "x-permitted-roles": [
          "Owner",
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "organization_administration_forbidden",
            "organization_enabled_change_forbidden",
            "csrf_required"
          ],
          "404": [
            "organization_not_found"
          ],
          "422": [
            "invalid_organization_request"
          ]
        }
      }
    },
    "/api/v1/organizations/{id}/disable": {
      "post": {
        "operationId": "disableOrganization",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "examples": {
                  "example": {
                    "summary": "Error example",
                    "value": {
                      "code": "INVALID_TOKEN",
                      "detail": "Request failed",
                      "field_errors": null,
                      "request_id": "request-id",
                      "status": 400,
                      "title": "Bad Request",
                      "type": "about:blank"
                    }
                  }
                },
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ],
        "summary": "Disable an organization",
        "x-csrf-required": true,
        "x-permitted-roles": [
          "Superadmin"
        ],
        "x-stable-error-codes": {
          "401": [
            "invalid_session"
          ],
          "403": [
            "organization_administration_forbidden",
            "csrf_required"
          ],
          "404": [
            "organization_not_found"
          ]
        }
      }
    },
...
  },
...
}